All living entities are a Projectile source
