+++ src/main/java/org/spongepowered/common/entity/projectile/ProjectileLauncher.java
@@ -26,10 +26,6 @@ package org.spongepowered.common.entity.projectile;
 
 import com.flowpowered.math.vector.Vector3d;
 import com.google.common.collect.Maps;
-import net.minecraft.block.BlockDispenser;
-import net.minecraft.block.BlockSourceImpl;
-import net.minecraft.block.state.IBlockState;
-import net.minecraft.dispenser.BehaviorDefaultDispenseItem;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.item.EntityArmorStand;
 import net.minecraft.entity.item.EntityEnderPearl;
@@ -49,7 +45,6 @@ import net.minecraft.entity.projectile.EntityThrowable;
 import net.minecraft.entity.projectile.EntityTippedArrow;
 import net.minecraft.entity.projectile.EntityWitherSkull;
 import net.minecraft.init.Items;
-import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.tileentity.TileEntityDispenser;
 import net.minecraft.util.EnumFacing;
@@ -80,99 +75,17 @@ import org.spongepowered.api.event.cause.Cause;
 import org.spongepowered.api.event.cause.entity.spawn.SpawnCause;
 import org.spongepowered.api.event.entity.projectile.LaunchProjectileEvent;
 import org.spongepowered.api.world.Location;
-import org.spongepowered.api.world.World;
 import org.spongepowered.api.world.extent.Extent;
 import org.spongepowered.common.SpongeImpl;
 import org.spongepowered.common.event.SpongeCommonEventFactory;
 import org.spongepowered.common.registry.type.entity.EntityTypeRegistryModule;
 import org.spongepowered.common.registry.type.event.InternalSpawnTypes;
 
-import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Random;
 
 public class ProjectileLauncher {
-
-    public interface ProjectileSourceLogic<T extends ProjectileSource> {
-
-        <P extends Projectile> Optional<P> launch(ProjectileLogic<P> logic, T source, Class<P> projectileClass, Object... args);
-
-    }
-
-    public interface ProjectileLogic<T extends Projectile> {
-
-        Optional<T> launch(ProjectileSource source);
-
-        default Optional<T> createProjectile(ProjectileSource source, Class<T> projectileClass, Location<?> loc) {
-            return defaultLaunch(source, projectileClass, loc);
-        }
-    }
-
-    public static class SimpleEntityLaunchLogic<P extends Projectile> implements ProjectileLogic<P> {
-
-        protected final Class<P> projectileClass;
-
-        public SimpleEntityLaunchLogic(Class<P> projectileClass) {
-            this.projectileClass = projectileClass;
-        }
-
-        @Override
-        public Optional<P> launch(ProjectileSource source) {
-            if (!(source instanceof Entity)) {
-                return Optional.empty();
-            }
-            Location<World> loc = ((Entity) source).getLocation().add(0, ((net.minecraft.entity.Entity) source).height / 2, 0);
-            Optional<P> projectile;
-            if (source instanceof EntityLivingBase) {
-                projectile = createProjectile((EntityLivingBase) source, loc);
-            } else {
-                projectile = createProjectile(source, this.projectileClass, loc);
-            }
-            return projectile;
-        }
-
-        protected Optional<P> createProjectile(EntityLivingBase source, Location<?> loc) {
-            return createProjectile((ProjectileSource) source, this.projectileClass, loc);
-        }
-    }
-
-    public static class SimpleDispenserLaunchLogic<P extends Projectile> extends SimpleEntityLaunchLogic<P> {
-
-        public SimpleDispenserLaunchLogic(Class<P> projectileClass) {
-            super(projectileClass);
-        }
-
-        @Override
-        public Optional<P> launch(ProjectileSource source) {
-            Optional<P> ret = super.launch(source);
-            if (ret.isPresent()) {
-                return ret;
-            }
-            if (source instanceof Dispenser) {
-                return getSourceLogic(Dispenser.class).launch(this, (Dispenser) source, this.projectileClass);
-            }
-            return ret;
-        }
-    }
-
-    public static class SimpleItemLaunchLogic<P extends Projectile> extends SimpleEntityLaunchLogic<P> {
-
-        private final Item item;
-
-        public SimpleItemLaunchLogic(Class<P> projectileClass, Item item) {
-            super(projectileClass);
-            this.item = item;
-        }
-
-        @Override
-        public Optional<P> launch(ProjectileSource source) {
-            if (source instanceof TileEntityDispenser) {
-                return getSourceLogic(Dispenser.class).launch(this, (Dispenser) source, this.projectileClass, this.item);
-            }
-            return super.launch(source);
-        }
-    }
 
     private static final Map<Class<? extends Projectile>, ProjectileLogic<?>> projectileLogic = Maps.newHashMap();
     private static final Map<Class<? extends ProjectileSource>, ProjectileSourceLogic<?>> projectileSourceLogic = Maps.newHashMap();
@@ -183,6 +96,12 @@ public class ProjectileLauncher {
             return Optional.empty();
         }
         Optional<T> projectile = logic.launch(source);
+        projectile.ifPresent(t -> {
+            if (vel != null) {
+                t.offer(Keys.VELOCITY, vel);
+            }
+            t.setShooter(source);
+        });
         if (projectile.isPresent()) {
             if (vel != null) {
                 projectile.get().offer(Keys.VELOCITY, vel);
@@ -215,12 +134,12 @@ public class ProjectileLauncher {
     }
 
     @SuppressWarnings("unchecked")
-    public static <T extends ProjectileSource> ProjectileSourceLogic<T> getSourceLogic(Class<T> sourceClass) {
+    static <T extends ProjectileSource> ProjectileSourceLogic<T> getSourceLogic(Class<T> sourceClass) {
         return (ProjectileSourceLogic<T>) projectileSourceLogic.get(sourceClass);
     }
 
     @SuppressWarnings("unchecked")
-    public static <T extends Projectile> ProjectileLogic<T> getLogic(Class<T> sourceClass) {
+    private static <T extends Projectile> ProjectileLogic<T> getLogic(Class<T> sourceClass) {
         // If a concrete class is handed to us, find the API interface
         if (!sourceClass.isInterface() && net.minecraft.entity.Entity.class.isAssignableFrom(sourceClass)) {
             for (Class<?> iface : sourceClass.getInterfaces()) {
@@ -240,9 +159,6 @@ public class ProjectileLauncher {
             return Optional.empty();
         }
         Entity projectile = loc.getExtent().createEntity(opType.get(), loc.getPosition());
-        if (projectile == null) {
-            return Optional.empty();
-        }
         if (projectile instanceof EntityThrowable) {
             configureThrowable((EntityThrowable) projectile);
         }
@@ -250,12 +166,12 @@ public class ProjectileLauncher {
     }
 
     // Internal
-    static Cause createCause(Object source) {
+    private static Cause createCause(Object source) {
         return Cause.source(SpawnCause.builder().type(InternalSpawnTypes.PROJECTILE).build())
                 .named("ProjectileSource", source).build();
     }
 
-    static <P extends Projectile> Optional<P> doLaunch(Extent extent, P projectile, Cause cause) {
+    private static <P extends Projectile> Optional<P> doLaunch(Extent extent, P projectile, Cause cause) {
         SpongeCommonEventFactory.checkSpawnEvent(projectile, cause);
         LaunchProjectileEvent event = SpongeEventFactory.createLaunchProjectileEvent(cause, projectile);
         SpongeImpl.getGame().getEventManager().post(event);
@@ -265,50 +181,7 @@ public class ProjectileLauncher {
         return Optional.empty();
     }
 
-    private static class DispenserSourceLogic implements ProjectileSourceLogic<Dispenser> {
-
-        DispenserSourceLogic() {
-        }
-
-        @Override
-        public <P extends Projectile> Optional<P> launch(ProjectileLogic<P> logic, Dispenser source, Class<P> projectileClass, Object... args) {
-            if (args.length == 1 && args[0] instanceof Item) {
-                return launch((TileEntityDispenser) source, projectileClass, (Item) args[0]);
-            }
-            Optional<P> projectile = logic.createProjectile(source, projectileClass, source.getLocation());
-            if (projectile.isPresent()) {
-                EnumFacing enumfacing = getFacing((TileEntityDispenser) source);
-                net.minecraft.entity.Entity projectileEntity = (net.minecraft.entity.Entity) projectile.get();
-                projectileEntity.motionX = enumfacing.getFrontOffsetX();
-                projectileEntity.motionY = enumfacing.getFrontOffsetY() + 0.1F;
-                projectileEntity.motionZ = enumfacing.getFrontOffsetZ();
-            }
-            return projectile;
-        }
-
-        public static EnumFacing getFacing(TileEntityDispenser dispenser) {
-            IBlockState state = dispenser.getWorld().getBlockState(dispenser.getPos());
-            return state.getValue(BlockDispenser.FACING);
-        }
-
-        @SuppressWarnings("unchecked")
-        private <P extends Projectile> Optional<P> launch(TileEntityDispenser dispenser, Class<P> projectileClass, Item item) {
-            BehaviorDefaultDispenseItem behavior = (BehaviorDefaultDispenseItem) BlockDispenser.DISPENSE_BEHAVIOR_REGISTRY.getObject(item);
-            List<net.minecraft.entity.Entity> entityList = dispenser.getWorld().loadedEntityList;
-            int numEntities = entityList.size();
-            behavior.dispense(new BlockSourceImpl(dispenser.getWorld(), dispenser.getPos()), new ItemStack(item));
-            // Hack - get the projectile that was spawned from dispense()
-            for (int i = entityList.size() - 1; i >= numEntities; i--) {
-                if (projectileClass.isInstance(entityList.get(i))) {
-                    return Optional.of((P) entityList.get(i));
-                }
-            }
-            return Optional.empty();
-        }
-    }
-
     static {
-
         registerProjectileSourceLogic(Dispenser.class, new DispenserSourceLogic());
 
         registerProjectileLogic(TippedArrow.class, new SimpleItemLaunchLogic<TippedArrow>(TippedArrow.class, Items.ARROW) {
